move src/containers/trends/ into appropriate sub-folders of src/components/trends/
