+++ src/index.js
@@ -25,7 +25,7 @@ import RangeSelect from './components/trends/cbg/RangeSelect';
 import TwoOptionToggle from './components/common/controls/TwoOptionToggle';
 
 import PumpSettingsContainer from './containers/settings/PumpSettingsContainer';
-import TrendsContainer from './containers/trends/TrendsContainer';
+import TrendsContainer from './components/trends/common/TrendsContainer';
 
 import vizReducer from './redux/reducers/';
